+++ src/common/colors.ts
@@ -2,6 +2,7 @@ const APP_COLORS = {
     white: '#ffffff',
     gray: '#A9A9A9',
     lightGray: '#C8C7CD',
+    darkGray: '#0000004D',
     red: '#dd2c00',
     black: '#000000',
     green: '#60bf2e',