+++ src/common/icons.ts
@@ -1,3 +1,5 @@
 export default class APP_ICONS {
     static plus = require('../../assets/plus.png')
+
+    static minus = require('../../assets/minus.png')
 }